front: move rollingStock helpers to appropriate folders
Some helpers were specific to RollingStockEditor and should be moved in the relevant folder.
Helpers have been distributed in new files, such as the corresponding unit tests.

I only moved code around.
