+++ front/src/modules/rollingStock/components/RollingStockEditor/helpers/__tests__/curves.spec.ts
@@ -1,18 +1,12 @@
 import { describe, it, expect } from 'vitest';
 
-import {
-  createEmptyCurve,
-  createEmptyCurves,
-  orderElectricalProfils,
-  sortSelectedModeCurves,
-} from 'modules/rollingStock/helpers/utils';
+import { sortSelectedModeCurves } from '../curves';
+import { createEmptyCurves, createEmptyCurve } from '../defaultValues';
 
 const emptyEffortCurves = createEmptyCurves('1500V', ['STANDARD', 'AIR_CONDITIONING', 'HEATING']);
 const newCurveWithAllCond = createEmptyCurve('STANDARD', 'A1', 'C1US');
 const newCurveWithOnlyPR = createEmptyCurve('STANDARD', null, 'C1US');
 const newCurveWithOnlyEP = createEmptyCurve('STANDARD', 'B');
-
-const electricalProfilesList1500 = ['B', 'C', 'A', null, 'O'];
 
 describe('sortSelectedModeCurves', () => {
   it('should order the curves list and put the newCurve with all conditions at the beginning of the list', () => {
@@ -57,15 +51,6 @@ describe('sortSelectedModeCurves', () => {
       newCurveWithOnlyEP,
       ...emptyEffortCurves.curves,
     ];
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe('orderElectricalProfils', () => {
-  it('should order the list of electrical profiles', () => {
-    const result = orderElectricalProfils(electricalProfilesList1500, '1500V');
-    const expected = [null, 'O', 'A', 'B', 'C'];
 
     expect(result).toEqual(expected);
   });
