+++ front/src/modules/rollingStock/consts.ts
@@ -1,7 +1,5 @@
 import type { Comfort, TrainMainCategory } from 'common/api/osrdEditoastApi';
 
-import type { MultiUnit } from './types';
-
 export const THERMAL_TRACTION_IDENTIFIER = 'thermal';
 export const STANDARD_COMFORT_LEVEL: Comfort = 'STANDARD';
 
@@ -26,21 +24,4 @@ export const TrainMainCategoryDict: Record<TrainMainCategory, TrainMainCategory>
   TRAM_TRAIN: 'TRAM_TRAIN',
   TOURISTIC_TRAIN: 'TOURISTIC_TRAIN',
   WORK_TRAIN: 'WORK_TRAIN',
-};
-
-export const CONVERSION_FACTORS_SCHEMA: Partial<
-  Record<MultiUnit, Partial<Record<MultiUnit, number>>>
-> = {
-  t: { kg: 1000 },
-  kg: { t: 1 / 1000 },
-  'km/h': { 'm/s': 1 / 3.6 },
-  'm/s': { 'km/h': 3.6 },
-  N: { kN: 1 / 1000 },
-  kN: { N: 1000 },
-  'N/(m/s)': { 'N/(km/h)': 1 / 3.6, 'kN/(km/h)': 1 / (1000 * 3.6) },
-  'N/(km/h)': { 'N/(m/s)': 3.6, 'kN/(km/h)': 1 / 1000 },
-  'kN/(km/h)': { 'N/(m/s)': 1000 * 3.6, 'N/(km/h)': 1000 },
-  'N/(m/s)²': { 'N/(km/h)²': 1 / 3.6 ** 2, 'kN/(km/h)²': 1 / (1000 * 3.6 ** 2) },
-  'N/(km/h)²': { 'N/(m/s)²': 3.6 ** 2, 'kN/(km/h)²': 1 / 1000 },
-  'kN/(km/h)²': { 'N/(m/s)²': 1000 * 3.6 ** 2, 'N/(km/h)²': 1000 },
 };
